add Fact class in its own file. lint main files.
